Refactor countdown formatting and improve token/network utils
Extracted countdown formatting logic to a new utils/time.ts file with a reusable formatDurationShort function. Updated LiquidityRequestsCard to use this utility. Improved useAccountFtBalance to memoize decimals and avoid redundant calculations. Enhanced network explorer URL helper to throw on unsupported networks. Added detailed comments to USDC mainnet token ID regarding implicit accounts.
